requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.0
cryptography>=41.0.0
msal>=1.24.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
websockets>=12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
